Fix VFS transport issues
Add capability to specify url parameters for MoveAfterProcess
and MoveAfterFailure parameters in vfs proxy.
Fixes wso2/product-ei#3088

Set sftp file options regardless of FileURI scheme since the
sftp options are required when one of FileURI,MoveAfterProcess
or ActionAfterFailure is set to use sftp it is incorrect not to set
sftp file options based only on the scheme of FileURI.
Fixes wso2/product-ei#3035

Decrypt parameters if the encrypted content is specified in the
{wso2:vault-decrypt('param value')} format. Applied to all the
parameters specified in a vfs proxy. Since the decrypted parameter
values are kept in the ParameterInclude instance it self they can be
used later without having to be decrypted again.
Fixes wso2/product-ei#3125, Fixes wso2/product-ei#3031